Move the sig-instrumentation test to a dedicated folder
