+++ test/e2e/BUILD
@@ -12,7 +12,6 @@ go_test(
     name = "go_default_test",
     srcs = [
         "e2e_test.go",
-        "metrics_grabber_test.go",
         "taints_test.go",
     ],
     library = ":go_default_library",
@@ -26,14 +25,12 @@ go_test(
         "//test/e2e/kubectl:go_default_library",
         "//test/e2e/lifecycle:go_default_library",
         "//test/e2e/lifecycle/bootstrap:go_default_library",
-        "//test/e2e/metrics:go_default_library",
         "//test/e2e/network:go_default_library",
         "//test/e2e/scalability:go_default_library",
         "//test/e2e/scheduling:go_default_library",
         "//test/e2e/storage:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
-        "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/github.com/stretchr/testify/assert:go_default_library",
         "//vendor/k8s.io/api/core/v1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
@@ -85,9 +82,9 @@ go_library(
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
         "//test/e2e/framework/ginkgowrapper:go_default_library",
+        "//test/e2e/framework/metrics:go_default_library",
         "//test/e2e/generated:go_default_library",
         "//test/e2e/manifest:go_default_library",
-        "//test/e2e/metrics:go_default_library",
         "//test/e2e_federation:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
@@ -155,7 +152,6 @@ filegroup(
         "//test/e2e/kubectl:all-srcs",
         "//test/e2e/lifecycle:all-srcs",
         "//test/e2e/manifest:all-srcs",
-        "//test/e2e/metrics:all-srcs",
         "//test/e2e/network:all-srcs",
         "//test/e2e/node:all-srcs",
         "//test/e2e/perftype:all-srcs",
