+++ test/e2e/framework/BUILD
@@ -79,8 +79,8 @@ go_library(
         "//plugin/pkg/scheduler/algorithm/predicates:go_default_library",
         "//plugin/pkg/scheduler/schedulercache:go_default_library",
         "//test/e2e/framework/ginkgowrapper:go_default_library",
+        "//test/e2e/framework/metrics:go_default_library",
         "//test/e2e/manifest:go_default_library",
-        "//test/e2e/metrics:go_default_library",
         "//test/e2e/perftype:go_default_library",
         "//test/utils:go_default_library",
         "//vendor/github.com/aws/aws-sdk-go/aws:go_default_library",
@@ -152,6 +152,7 @@ filegroup(
     srcs = [
         ":package-srcs",
         "//test/e2e/framework/ginkgowrapper:all-srcs",
+        "//test/e2e/framework/metrics:all-srcs",
     ],
     tags = ["automanaged"],
 )
