+++ test/e2e/framework/metrics/BUILD
