+++ test/e2e/framework/metrics/controller_manager_metrics.go
