+++ test/e2e/instrumentation/monitoring/BUILD
@@ -12,15 +12,18 @@ go_library(
     srcs = [
         "cadvisor.go",
         "influxdb.go",
+        "metrics_grabber.go",
         "stackdriver.go",
     ],
     tags = ["automanaged"],
     deps = [
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/framework/metrics:go_default_library",
         "//test/e2e/instrumentation/common:go_default_library",
         "//vendor/github.com/influxdata/influxdb/client/v2:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
+        "//vendor/github.com/onsi/gomega:go_default_library",
         "//vendor/golang.org/x/oauth2/google:go_default_library",
         "//vendor/google.golang.org/api/monitoring/v3:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
