+++ test/e2e_node/BUILD
@@ -36,7 +36,7 @@ go_library(
         "//pkg/util/procfs:go_default_library",
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
-        "//test/e2e/metrics:go_default_library",
+        "//test/e2e/framework/metrics:go_default_library",
         "//test/e2e/perftype:go_default_library",
         "//test/e2e_node/perftype:go_default_library",
         "//vendor/github.com/golang/glog:go_default_library",
@@ -114,7 +114,7 @@ go_test(
         "//pkg/security/apparmor:go_default_library",
         "//test/e2e/common:go_default_library",
         "//test/e2e/framework:go_default_library",
-        "//test/e2e/metrics:go_default_library",
+        "//test/e2e/framework/metrics:go_default_library",
         "//test/e2e_node/services:go_default_library",
         "//test/e2e_node/system:go_default_library",
         "//test/utils:go_default_library",
